bist: Move reader/writer to separate classes
